+++ Logging/SFA.DAS.NLog.Logger/SFA.DAS.NLog.Logger.csproj
@@ -54,8 +54,8 @@
     <Compile Include="IRequestContext.cs" />
     <Compile Include="NLogLogger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Web\CorrelationIdMessageHandler.cs" />
-    <Compile Include="Web\DelegatingHandler\CorrelationIdDelegatingHandler.cs" />
+    <Compile Include="Web\MessageHandlers\CorrelationIdMessageResponseHandler.cs" />
+    <Compile Include="Web\MessageHandlers\CorrelationIdMessageRequestHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
